+++ ts/components/game/Choices.tsx
@@ -6,10 +6,10 @@ import Option from "../../Option";
 import Card from "./Card";
 import {GameTag} from "../../enums";
 import {CardType} from "../../enums";
-import {EntityListProps} from "../../interfaces";
+import {EntityListProps, MulliganOracleProps} from "../../interfaces";
 import Choice from "../../Choice";
 
-interface ChoicesProps extends EntityListProps {
+interface ChoicesProps extends EntityListProps, MulliganOracleProps {
 	isMulligan: boolean;
 	choices: Immutable.Map<number, Choice>;
 }
@@ -37,7 +37,7 @@ class Choices extends EntityList<ChoicesProps> {
 		if(this.props.hideCards) {
 			entity = new Entity(entity.id, entity.getTags());
 		}
-		else if (!entity.cardId && this.props.cardOracle && this.props.cardOracle.has(+entity.id)) {
+		else if (!entity.cardId && this.props.cardOracle && this.props.cardOracle.has(entity.id)) {
 			let cardId = this.props.cardOracle.get(entity.id);
 			entity = new Entity(entity.id, entity.getTags(), cardId);
 			hidden = true;
@@ -49,14 +49,23 @@ class Choices extends EntityList<ChoicesProps> {
 			return null;
 		}
 
-		return (<Card entity={entity}
-			option={option}
-			optionCallback={this.props.optionCallback}
-			assetDirectory={this.props.assetDirectory}
-			cards={this.props.cards}
-			isHidden={hidden}
-			controller={this.props.controller}
-			cardArtDirectory={this.props.cardArtDirectory}
+		// mulligan cards
+		let mulligan = false;
+		if(this.props.isMulligan && this.props.mulliganOracle && this.props.mulliganOracle.get(entity.id) === true) {
+			mulligan = true;
+		}
+
+		return (
+			<Card
+				entity={entity}
+				option={option}
+				optionCallback={this.props.optionCallback}
+				assetDirectory={this.props.assetDirectory}
+				cards={this.props.cards}
+				isHidden={hidden}
+				controller={this.props.controller}
+				cardArtDirectory={this.props.cardArtDirectory}
+				mulligan={mulligan}
 			/>);
 	}
 }